+++ packages/rx_bloc_cli/lib/src/commands/create_command.dart
@@ -10,6 +10,7 @@ import '../templates/feature_widget_toolkit_bundle.dart';
 import '../templates/lib_auth_bundle.dart';
 import '../templates/lib_change_language_bundle.dart';
 import '../templates/lib_permissions_bundle.dart';
+import '../templates/lib_realtime_communication_bundle.dart';
 import '../templates/lib_router_bundle.dart';
 import '../templates/lib_social_logins_bundle.dart';
 import '../templates/patrol_integration_tests_bundle.dart';
@@ -80,6 +81,13 @@ class CreateCommand extends Command<int> {
         help: 'Enables Patrol integration tests for the project',
         allowed: ['true', 'false'],
         defaultsTo: 'false',
+      )
+      ..addOption(
+        _realtimeCommunicationString,
+        help: 'Enables realtime communication facilities like SSE, WebSocket '
+            'or gRPC',
+        allowed: ['none', 'sse'],
+        defaultsTo: 'none',
       );
   }
 
@@ -94,6 +102,7 @@ class CreateCommand extends Command<int> {
   final _socialLoginsString = 'enable-social-logins';
   final _changeLanguageString = 'enable-change-language';
   final _patrolTestsString = 'enable-patrol';
+  final _realtimeCommunicationString = 'realtime-communication';
 
   /// bundles
   final _counterBundle = featureCounterBundle;
@@ -105,6 +114,7 @@ class CreateCommand extends Command<int> {
   final _libSocialLoginsBundle = libSocialLoginsBundle;
   final _libChangeLanguageBundle = libChangeLanguageBundle;
   final _patrolIntegrationTestsBundle = patrolIntegrationTestsBundle;
+  final _libRealtimeCommunicationBundle = libRealtimeCommunicationBundle;
 
   final Logger _logger;
   final MasonBundle _bundle;
@@ -234,6 +244,11 @@ class CreateCommand extends Command<int> {
       _bundle.files.addAll(_patrolIntegrationTestsBundle.files);
     }
 
+    if (arguments.realtimeCommunicationType !=
+        _RealtimeCommunicationType.none) {
+      _bundle.files.addAll(_libRealtimeCommunicationBundle.files);
+    }
+
     //Add lib_route to _bundle
     _bundle.files.addAll(_libRouterBundle.files);
     //Add lib_permissions to _bundle
@@ -259,6 +274,8 @@ class CreateCommand extends Command<int> {
         'enable_social_logins': arguments.enableSocialLogins,
         'enable_change_language': arguments.enableChangeLanguage,
         'enable_patrol': arguments.enablePatrolTests,
+        'realtime_communication': arguments.realtimeCommunicationType !=
+            _RealtimeCommunicationType.none,
       },
     );
 
@@ -288,6 +305,7 @@ class CreateCommand extends Command<int> {
       enableSocialLogins: _parseEnableSocialLogins(arguments),
       enableChangeLanguage: _parseEnableChangeLanguage(arguments),
       enablePatrolTests: _parseEnablePatrolTests(arguments),
+      realtimeCommunicationType: _parseRealtimeCommunicationType(arguments),
     );
   }
 
@@ -352,6 +370,24 @@ class CreateCommand extends Command<int> {
   bool _parseEnableSocialLogins(ArgResults arguments) {
     final socialLoginsEnabled = arguments[_socialLoginsString];
     return socialLoginsEnabled.toLowerCase() == 'true';
+  }
+
+  _RealtimeCommunicationType _parseRealtimeCommunicationType(arguments) {
+    final type = arguments[_realtimeCommunicationString] as String;
+
+    switch (type.toLowerCase()) {
+      case 'sse':
+        return _RealtimeCommunicationType.sse;
+      case 'websocket':
+        return _RealtimeCommunicationType.websocket;
+      case 'grpc':
+        return _RealtimeCommunicationType.grpc;
+      case 'none':
+        return _RealtimeCommunicationType.none;
+    }
+
+    throw UsageException(
+        'Unexpected value for parameter $_realtimeCommunicationString.', usage);
   }
 
   /// endregion
@@ -469,6 +505,7 @@ class _CreateCommandArguments {
     required this.enableSocialLogins,
     required this.enableChangeLanguage,
     required this.enablePatrolTests,
+    required this.realtimeCommunicationType,
   });
 
   final String projectName;
@@ -481,4 +518,7 @@ class _CreateCommandArguments {
   final bool enableSocialLogins;
   final bool enableChangeLanguage;
   final bool enablePatrolTests;
+  final _RealtimeCommunicationType realtimeCommunicationType;
 }
+
+enum _RealtimeCommunicationType { none, sse, websocket, grpc }