+++ packages/rx_bloc_cli/mason_templates/bricks/lib_realtime_communication/__brick__/lib/base/models/response_models/sse_message_model.dart
@@ -1,0 +1,13 @@
+class SseMessageModel {
+  const SseMessageModel({
+    this.id,
+    required this.event,
+    required this.data,
+    this.retry,
+  });
+
+  final String? id;
+  final String event;
+  final String data;
+  final int? retry;
+}