+++ packages/rx_bloc_cli/CHANGELOG.md
@@ -1,8 +1,6 @@
-## [x.x.x]
-* Added SSE communication facilities template
-
 ## [2.6.0]
 * Added parameter `enable-patrol` for enabling patrol package for integration tests
+* Added SSE communication facilities template
 * Updated to Dart 3.0
 
 ## [2.5.0]